create my_app file for MyApp widget from main.dart
